renamed boolean target to nominal taret as suggested by ZeCarlosCoutinho
