+++ pysubgroup/__init__.py
@@ -3,7 +3,7 @@ from pysubgroup.algorithms import *
 from pysubgroup.measures import *
 from pysubgroup.utils import *
 
-from pysubgroup.boolean_target import *
+from pysubgroup.nominal_target import *
 from pysubgroup.numeric_target import *
 from pysubgroup.fi_target import *
